telegrambot_billing/requirements.txt
aws-cdk-lib==2.212.0
constructs>=10.0.0,<11.0.0
aws-cdk.aws-lambda-python-alpha==2.212.0a0
